Create component for table
